Move scripts under scripts dir
Also made scripts runnable from anywhere
