+++ flavors/grapheneos/extract-upstream-params.sh
@@ -8,6 +8,8 @@
 
 set -euo pipefail
 
+cd "$(dirname "${BASH_SOURCE[0]}")"
+
 DEVICE=sunfish
 CHANNEL=beta
