+++ modules/apv-lib-check.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 # SPDX-FileCopyrightText: 2020 Daniel Fullmer and robotnix contributors
 # SPDX-License-Identifier: MIT
+
 set -euo pipefail
 
 BUILT=$1
