+++ scripts/check.sh
@@ -2,6 +2,8 @@
 
 set -euo pipefail
 
+cd "$(dirname "${BASH_SOURCE[0]}")/.."
+
 mypy --exclude apks/chromium .
 flake8 --exclude apks/chromium .
 shellcheck ./*.sh flavors/*/*.sh modules/pixel/update.sh
