+++ scripts/mk-repo-file.py
@@ -2,7 +2,7 @@
 # SPDX-FileCopyrightText: 2020 Daniel Fullmer and robotnix contributors
 # SPDX-License-Identifier: MIT
 
-from typing import Any, Optional, Dict, List, Tuple
+from typing import Optional, Dict, List, Tuple
 from enum import Enum
 
 import argparse
@@ -11,6 +11,8 @@ import os
 import re
 import subprocess
 import tempfile
+
+from robotnix_common import save, checkout_git, ls_remote
 
 REPO_FLAGS = [
     "--quiet",
@@ -31,26 +33,6 @@ class ManifestRefType(Enum):
 revHashes: Dict[Tuple[str, bool], str] = {}  # (rev, fetch_submodules) -> sha256hash
 revTrees: Dict[str, str] = {}           # rev -> treeHash
 treeHashes: Dict[Tuple[str, bool], str] = {}  # (treeHash, fetch_submodules) -> sha256hash
-
-
-def save(filename: str, data: Any) -> None:
-    open(filename, 'w').write(json.dumps(data, sort_keys=True, indent=2, separators=(',', ': ')))
-
-
-def checkout_git(url: str, rev: str, fetch_submodules: bool = False) -> Any:
-    print("Checking out %s %s" % (url, rev))
-    args = ["nix-prefetch-git", "--url", url, "--rev", rev]
-    if fetch_submodules:
-        args.append("--fetch-submodules")
-    json_text = subprocess.check_output(args).decode()
-    return json.loads(json_text)
-
-
-def ls_remote(url: str, rev: str) -> str:
-    remote_info = subprocess.check_output(["git", "ls-remote", url, rev]).decode()
-    remote_rev = remote_info.split('\t')[0]
-    assert remote_rev != ""
-    return remote_rev
 
 
 def make_repo_file(url: str, ref: str, filename: str, ref_type: ManifestRefType,
